requirements.txt
pyrogram==1.2.9
tgcrypto
python-decouple